+++ Sources/CanvasKit/Mask/Mask.swift
@@ -188,7 +188,7 @@ public final class Mask: LayerProtocol, @unchecked Sendable {
     ///
     /// - Parameters:
     ///   - threshold: If pixel value is greater than `threshold`, the pixel is set to 255.
-    public func quantized(threshold: Float16) async throws -> Mask {
+    public func quantized(threshold: Float) async throws -> Mask {
         let newTexture = Mask.makeTexture(width: self.width, height: self.height)
         
         try await MetalFunction(name: "mask_quantize", bundle: .module)
@@ -229,7 +229,7 @@ public final class Mask: LayerProtocol, @unchecked Sendable {
     ///
     /// - Parameters:
     ///   - uint8: The mask value. A value of zero would indicate not selected, while any none-zero value would indicate the given pixel is selected.
-    public convenience init(repeating float: Float16 = 0, width: Int, height: Int, context: MetalContext) async throws {
+    public convenience init(repeating float: Float = 0, width: Int, height: Int, context: MetalContext) async throws {
         let texture = Mask.makeTexture(width: width, height: height)
         
         try await MetalFunction(name: "mask_fill_with", bundle: .module)
